implement a metadata_service class
