+++ tests/unit_tests/services/test_file_system_service.py
@@ -1,11 +1,10 @@
 
-import os
 import shutil
 import tempfile
 import unittest
 
 from delivery.services.file_system_service import FileSystemService
-from tests import test_utils
+
 
 class TestFileSystemService(unittest.TestCase):
 
@@ -35,8 +34,3 @@ class TestFileSystemService(unittest.TestCase):
             sorted(self.files),
             sorted(list(FileSystemService().list_files_recursively(self.rootdir)))
         )
-
-    def test_extract_samplesheet_data(self):
-        runfolder = test_utils.unorganised_runfolder(self.rootdir)
-        samplesheet_file, samplesheet_data = test_utils.samplesheet_file_from_runfolder(runfolder)
-        self.assertListEqual(samplesheet_data, FileSystemService.extract_samplesheet_data(samplesheet_file))
